pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "odinslist"
version = "0.2.0"
description = "Automated comic book cataloging using vision language models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "openai>=1.0",
    "Pillow>=9.0",
    "curl-cffi>=0.5",
    "tomli-w>=1.0",
    "tomli>=2.0; python_version < \"3.11\"",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
]

[project.scripts]
odinslist = "odinslist.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
